pyproject.toml
[tool]
[tool.poetry]
name="Vibhi-Hime"
version="0.0.1"
description=""
authors = ["ATCtech"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
python-decouple = "^3.3"
praw = "^7.1.0"
wikipedia = "^1.4.0"
disputils = "^0.1.3"
youtube_dl = "^2020.7.28"
PyNaCl = "^1.4.0"
discord = "^1.0.1"
async-timeout = "^3.0.1"
Flask = "^1.1.2"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"